+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 1.4.5
+
+- Add swap costs to home page
+- Fix error reporting
+
 ## 1.4.4
 
 - Fix panic in v1.4.3